+++ src/Cwaff-Guns/Femtobyte.cs
@@ -68,7 +68,7 @@ public class Femtobyte : CwaffGun
                 enemyGuid = guid,
                 data      = new(
                     prefabName: enemy.ActorName,
-                    displayName: enemy.ActorName,
+                    displayName: guid.AmmonomiconName(),
                     prefab: enemy.gameObject),
             };
         }
@@ -176,7 +176,7 @@ public class Femtobyte : CwaffGun
         if (enemy.healthHaver is not HealthHaver hh || hh.IsDead || hh.IsBoss || hh.IsSubboss)
             return false;
         this._lastEnemyKilled = enemy.EnemyGuid;
-        this._lastEnemyName = EnemyDatabase.GetOrLoadByGuid(this._lastEnemyKilled).ActorName;
+        this._lastEnemyName = this._lastEnemyKilled.AmmonomiconName();
         this._displayNameDirty = true;
         if (this.PlayerOwner.HasSynergy(Synergy.MASTERY_FEMTOBYTE))
             if (DigitizedObject.FromEnemyGuid(enemy.EnemyGuid) is DigitizedObject d)